Refactor Calendar
